BLUEBUTTON-765: Moving PagingArguments into its own class
PagingArguments will be reused by the Patient and Coverage resource so
moving it to a separate class to better maintain it.
